* Corrected mistake with packet checksumming whereby the checksum was included into the length when calculating it on packet receive. * Further C++ conversion of HashData module, adding constructors and operator ==. * Replaced SMMU's W_LumpCheckSum algorithm with standard CRC32 courtesy of HashData class.

+++ source/d_net.cpp
@@ -152,7 +152,7 @@ void ResetNet(void)
 //
 static void HSendPacket(int node, int flags)
 {
-   netbuffer->checksum = flags;
+   netbuffer->checksum = (uint32_t)flags;
    
    if(!node)
    {
